backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
databricks-sql-connector==3.0.0
numpy<2.0

